refactor schema files
