+++ schema/base.graphql
@@ -1,1 +1,3 @@
 scalar Time
+type Query
+type Mutation
