--- schema/mutation.graphql
@@ -1,11 +1,0 @@
-type Mutation {
-    # users
-    createUser(input: CreateUserInput): User!
-    updateUser(input: UpdateUserInput): User!
-    deleteUser(input: DeleteUserInput): User!
-
-    # todos
-    createTodo(input: CreateTodoInput): Todo!
-    updateTodo(input: UpdateTodoInput): Todo!
-    deleteTodo(input: DeleteTodoInput): Todo!
-}
